chore: regroup internal packages to have a more clean structure under SDK.
