+++ sdk/instrumentation/google.golang.org/grpc/internal/helloworld/helloworld.proto
